+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Added
 - DeviceDao
     - deleteDeviceByScadaId: delete devices belongs to scadaId
+- Add configUploaded property
 
 ## Changed
 - device type is adjusted to String